# docs/openapi.yaml
openapi: 3.0.1
info:
  title: 'EventXperience Backend API'
  description: 'Generated API reference for the EventXperience backend. Responses use the shared ApiResponse envelope unless noted otherwise.'
  version: v1
paths:
...
  /api/events:
    get:
      tags:
        - events
      operationId: Events_GetEvents
      parameters:
        - name: search
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: isPrivate
          in: query
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: category
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: tags
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: city
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: lat
          in: query
          schema:
            '$ref': '#/components/schemas/Double4'
        - name: lng
          in: query
          schema:
            '$ref': '#/components/schemas/Double4'
        - name: radiusKm
          in: query
          schema:
            '$ref': '#/components/schemas/Double4'
        - name: sortBy
          in: query
          schema:
            type: integer
            default: 0
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
...
  '/api/events/{eventId}/invitations':
    post:
      tags:
        - events
      operationId: EventInvitation_CreateInvitations
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateEventInvitationsRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/CreateEventInvitationsRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/CreateEventInvitationsRequest'
        required: true
      responses:
        201:
          description: Created
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    get:
      tags:
        - events
      operationId: EventInvitation_GetInvitations
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/events/{eventId}/invitations/{invitationId}/revoke':
    post:
      tags:
        - events
      operationId: EventInvitation_RevokeInvitation
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
        - name: invitationId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/events/{eventId}/invitation-links':
    post:
      tags:
        - events
      operationId: EventInvitation_CreateInvitationLink
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateEventInvitationLinkRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/CreateEventInvitationLinkRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/CreateEventInvitationLinkRequest'
        required: true
      responses:
        201:
          description: Created
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationLinkResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationLinkResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationLinkResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    get:
      tags:
        - events
      operationId: EventInvitation_GetInvitationLinks
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationLinkResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationLinkResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventInvitationLinkResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/events/{eventId}/invitation-links/{linkId}/revoke':
    post:
      tags:
        - events
      operationId: EventInvitation_RevokeInvitationLink
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
        - name: linkId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationLinkResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationLinkResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationLinkResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  /api/events/invitations/resolve:
    post:
      tags:
        - events
      operationId: EventInvitation_ResolveInvitation
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/ResolveEventInvitationRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/ResolveEventInvitationRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/ResolveEventInvitationRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationResolveResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationResolveResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationResolveResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
  /api/events/invitations/accept:
    post:
      tags:
        - events
      operationId: EventInvitation_AcceptInvitation
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/DecideEventInvitationRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/DecideEventInvitationRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/DecideEventInvitationRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/events/invitations/{invitationId}/accept':
    post:
      tags:
        - events
      operationId: EventInvitation_AcceptInvitationById
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  /api/events/invitations/decline:
    post:
      tags:
        - events
      operationId: EventInvitation_DeclineInvitation
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/DecideEventInvitationRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/DecideEventInvitationRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/DecideEventInvitationRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/events/invitations/{invitationId}/decline':
    post:
      tags:
        - events
      operationId: EventInvitation_DeclineInvitationById
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfEventInvitationDecisionResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
...
components:
  schemas:
...
    ApiResponseOfObject:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfPagedResponseOfEventResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/PagedResponseOfEventResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    Boolean:
      type: boolean
...
    CreateEventInvitationLinkRequest:
      type: object
      properties:
        maxRedemptions:
          '$ref': '#/components/schemas/Int322'
        expiresAt:
          '$ref': '#/components/schemas/DateTime2'
    CreateEventInvitationsRequest:
      type: object
      properties:
        userIds:
          '$ref': '#/components/schemas/ListOfInt322'
        emails:
          '$ref': '#/components/schemas/ListOfString2'
        expiresAt:
          '$ref': '#/components/schemas/DateTime'
...
    Double4:
      type: number
      format: double
      default: null
...
    Int32:
      type: integer
      format: int32
...
    Int325:
      type: integer
      format: int32
      default: 1
    Int326:
      type: integer
      format: int32
      default: 20
...
    String:
      type: string
...
    String7:
      type: string
      default: null
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Send the access token in the Authorization header as `Bearer {token}` for protected API routes.'
      scheme: bearer
      bearerFormat: JWT
...